refactor: move all protocol imports to a single file
