+++ src/presentation/controllers/signup.controller.spec.ts
@@ -1,6 +1,6 @@
 import { SignupController } from './signup.controller';
 import { MissingParamError, InvalidParamError, ServerError } from '../errors';
-import { EmailValidator } from '../protocols/email-validator';
+import { EmailValidator } from '../protocols';
 
 interface SutTypes {
   sut: SignupController;
