+++ src/presentation/controllers/signup.controller.ts
@@ -1,8 +1,6 @@
-import { HttpRequest, HttpResponse } from '../protocols/http';
+import { HttpRequest, HttpResponse, EmailValidator, Controller } from '../protocols';
 import { MissingParamError, InvalidParamError } from '../errors';
 import { badRequest, serverError } from '../helpers/http.helper';
-import { Controller } from '../protocols/controller';
-import { EmailValidator } from '../protocols/email-validator';
 
 export class SignupController implements Controller {
   private readonly emailValidator: EmailValidator;
